Finished tests and split up tests
