+++ tests/Unit Tests/test_create_course.py
@@ -1,8 +1,6 @@
-from django.test import TestCase, Client
-from django.urls import reverse
+from django.test import TestCase
 
 from myapp.models import Course, User
-from myapp.views import CreateCourse
 from myapp.Classes.supervisor import Supervisor
 
 
@@ -68,31 +66,4 @@ class TestNewCourse(TestCase):
         self.assertTrue(
             isinstance(
                 result, TypeError), "Expected a type error from create_course"
-        )
-
-
-class TestButtons(TestCase):
-    def setUp(self):
-        self.user = User.objects.create_user(
-            username='testuser', password='testpass'
-        )
-
-    def tearDown(self):
-        self.user.delete()
-
-    def test_course_list_page_accessible(self):
-        response = self.client.get(reverse("course_base"))
-        self.assertEqual(
-            response.status_code,
-            200,
-            "Expected status_code 200, got " + str(response.status_code),
-        )
-
-    def test_new_course_page_accessible(self):
-        self.client.login(username='testuser', password='testpass')
-        response = self.client.get(reverse("createcourse"))
-        self.assertEqual(
-            response.status_code,
-            200,
-            "Expected status_code 200, got " + str(response.status_code),
         )
